web-api/requirements.txt
aiormq
fastapi
uvicorn
python-socketio